+++ tensorflow_toolkit/textile_recognition/.gitignore
@@ -1,0 +1,2 @@
+model
+models
