+++ tensorflow_toolkit/textile_recognition/textile/model.py
@@ -17,6 +17,7 @@ import tensorflow as tf
 EMBEDDINGS_DIM = 256
 WEIGHT_DECAY = 0.0001
 
+
 def l2_normalized_embeddings(inputs):
     output = tf.keras.layers.GlobalAveragePooling2D()(inputs)
     output = tf.keras.layers.Reshape([1, 1, output.shape[-1]])(output)
@@ -26,6 +27,7 @@ def l2_normalized_embeddings(inputs):
     output = tf.nn.l2_normalize(output * 1000, axis=-1, epsilon=1e-13)
     return output
 
+
 def keras_applications_mobilenetv2(inputs, alpha=1.0):
     from tensorflow.keras.applications.mobilenet_v2 import MobileNetV2
 
@@ -34,6 +36,7 @@ def keras_applications_mobilenetv2(inputs, alpha=1.0):
 
     output = l2_normalized_embeddings(base_model.output)
     return tf.keras.Model(inputs, output)
+
 
 def keras_applications_resnet50(inputs):
     from tensorflow.keras.applications.resnet50 import ResNet50
